+++ substra/sdk/backends/local/backend.py
@@ -2,6 +2,7 @@ import copy
 import logging
 import os
 import shutil
+import typing
 import warnings
 from datetime import datetime
 from pathlib import Path
@@ -396,8 +397,6 @@ class Local(base.BaseBackend):
         return data_samples
 
     def _add_compute_plan(self, spec: schemas.ComputePlanSpec, spec_options: dict = None):
-        if spec.clean_models:
-            warnings.warn("'clean_models=True' is ignored on the local backend.")
         self._check_metadata(spec.metadata)
         # Get all the tuples and their dependencies
         tuple_graph, tuples = compute_plan_module.get_dependency_graph(spec)
@@ -421,7 +420,6 @@ class Local(base.BaseBackend):
             failed_count=0,
             done_count=0,
             owner=self._org_id,
-            delete_intermediary_models=spec.clean_models if spec.clean_models is not None else False,
         )
         compute_plan = self._db.add(compute_plan)
 
@@ -459,6 +457,8 @@ class Local(base.BaseBackend):
         )
 
         compute_plan_key, rank = self.__create_compute_plan_from_tuple(spec=spec, key=key, in_tuples=in_tuples)
+
+        _warn_on_transient_outputs(spec.outputs)
 
         # create model
         traintuple = models.Traintuple(
@@ -509,6 +509,8 @@ class Local(base.BaseBackend):
 
         assert len(spec.test_data_sample_keys) > 0
         worker = dataset.owner
+
+        _warn_on_transient_outputs(spec.outputs)
 
         if traintuple.compute_plan_key:
             compute_plan = self._db.get(schemas.Type.ComputePlan, traintuple.compute_plan_key)
@@ -571,6 +573,8 @@ class Local(base.BaseBackend):
             compute_plan.todo_count += 1
             compute_plan.status = models.Status.waiting
 
+        _warn_on_transient_outputs(spec.outputs)
+
         testtuple = models.Testtuple(
             test=models._Test(
                 data_manager_key=dataset_key,
@@ -627,6 +631,8 @@ class Local(base.BaseBackend):
             parent_task_keys.append(spec.in_head_model_key)
         if spec.in_trunk_model_key is not None:
             parent_task_keys.append(spec.in_trunk_model_key)
+
+        _warn_on_transient_outputs(spec.outputs)
 
         composite_traintuple = models.CompositeTraintuple(
             composite=models._Composite(
@@ -879,3 +885,9 @@ def _output_from_spec(outputs: Dict[str, schemas.ComputeTaskOutputSpec]) -> Dict
         # default isNone (= outputs are not computed yet)
         for identifier, output in outputs.items()
     }
+
+
+def _warn_on_transient_outputs(outputs: typing.Dict[str, schemas.ComputeTaskOutputSpec]):
+    for _, output in outputs.items():
+        if output.is_transient:
+            warnings.warn("`transient=True` is ignored in local mode")