+++ tests/data_factory.py
@@ -429,7 +429,7 @@ class AssetsFactory:
             metadata=metadata,
         )
 
-    def create_compute_plan(self, key=None, tag="", name="Test compute plan", clean_models=False, metadata=None):
+    def create_compute_plan(self, key=None, tag="", name="Test compute plan", metadata=None):
         return substra.sdk.schemas.ComputePlanSpec(
             key=key or random_uuid(),
             traintuples=[],
@@ -440,7 +440,6 @@ class AssetsFactory:
             tag=tag,
             name=name,
             metadata=metadata,
-            clean_models=clean_models,
         )
 
     def add_compute_plan_tuples(self, compute_plan):